pyproject.toml
[project]
name = "agentpark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai>=0.3.32",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "langgraph-cli[inmem]>=0.4.1",
    "matplotlib>=3.10.6",
    "pandas>=2.3.2",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
]
